+++ UniversityAdmission/UniversityAdmission.Server/Controllers/RequestController.cs
@@ -3,12 +3,14 @@ using UniversityAdmission.Domain.Dto;
 using UniversityAdmission.Domain.Dto.DtoForRequest;
 using UniversityAdmission.Domain.Repositories;
 using Microsoft.AspNetCore.Mvc;
+using AutoMapper;
+using UniversityAdmission.Server.Servicies;
 
 namespace UniversityAdmission.Server.Controllers;
 
 [Route("api/[controller]")]
 [ApiController]
-public class RequestController(IRepository<Applicant> repositoryApplicant, IRepository<ExamResult> repositoryExamResult, IRepository<Specialty> repositorySpecialty, IRepository<Application> repositoryApplication) : ControllerBase
+public class RequestController(IRepository<Applicant> repositoryApplicant, IRepository<ExamResult> repositoryExamResult, IRepository<Specialty> repositorySpecialty, IRepository<Application> repositoryApplication, IMapper mapper, IAnalayzerService AnalayzerService) : ControllerBase
 {
     /// <summary>
     /// Вывести информацию об абитуриентах из указанного города.
@@ -17,9 +19,7 @@ public class RequestController(IRepository<Applicant> repositoryApplicant, IRepo
     [HttpGet("applicant-by-city")]
     public ActionResult<IEnumerable<ApplicantDtoGet>> GetApplicantByCity(string сity = "Самара")
     {
-        var request = (from applicant in repositoryApplicant.GetAll()
-                       where applicant.City == сity
-                       select applicant);
+        var request = AnalayzerService.GetApplicantByCity(сity);
         return Ok(request);
     }
 
@@ -29,22 +29,8 @@ public class RequestController(IRepository<Applicant> repositoryApplicant, IRepo
     [HttpGet("applicant-older-ageLimit")]
     public ActionResult<IEnumerable<ApplicantDtoGet>> GetApplicantOlderValue(int ageLimit = 20)
     {
-        DateOnly currentDate = DateOnly.FromDateTime(DateTime.Now); // Текущая дата
-
-        var request = repositoryApplicant.GetAll()
-            .Where(applicant =>
-            {
-                int age = currentDate.Year - applicant.DateOfBirth.Year;
-                if ((currentDate.Month < applicant.DateOfBirth.Month) ||
-                (currentDate.Month == applicant.DateOfBirth.Month && currentDate.Day < applicant.DateOfBirth.Day))
-                {
-                    age--;
-                }
-
-                return age > ageLimit;
-            })
-            .OrderBy(applicant => applicant.FullName);
-        return Ok(request);
+        var request = AnalayzerService.GetApplicantOlderValue(ageLimit);
+        return Ok();
     }
 
     /// <summary>
@@ -55,24 +41,7 @@ public class RequestController(IRepository<Applicant> repositoryApplicant, IRepo
     [HttpGet("applicant-by-speciality")]
     public ActionResult<IEnumerable<ApplicantDtoWithExamResult>> GetApplicantsBySpecialtySortedByExamScores(string specialtyCode = "10.05.03D")
     {
-        var request = (from applicant in repositoryApplicant.GetAll()
-                       join application in repositoryApplication.GetAll() on applicant.IdApplicant equals application.ApplicantId
-                       join specialty in repositorySpecialty.GetAll() on application.SpecialtyId equals specialty.IdSpecialty
-                       join examResult in repositoryExamResult.GetAll() on applicant.IdApplicant equals examResult.ApplicantId
-                       where specialty.Code == specialtyCode
-                       group examResult by new
-                       {
-                           Applicant = applicant,
-                           SpecialtyName = specialty.Name
-                       } into applicantGroup
-                       let totalScore = applicantGroup.Sum(exam => exam.Score)
-                       orderby totalScore descending
-                       select new ApplicantDtoWithExamResult
-                       {
-                           Applicant = applicantGroup.Key.Applicant,
-                           SumExamResult = totalScore
-                       })
-                       .ToList();
+        var request = AnalayzerService.GetApplicantsBySpecialtySortedByExamScores(specialtyCode);
         return Ok(request);
     }
 
@@ -83,16 +52,7 @@ public class RequestController(IRepository<Applicant> repositoryApplicant, IRepo
     [HttpGet("applicant-count-by-speciality")]
     public ActionResult<IEnumerable<ApplicantCountForSpecialityDto>> GetCountApplicantToSpecialty()
     {
-        var request = (from application in repositoryApplication.GetAll()
-                       join specialty in repositorySpecialty.GetAll() on application.SpecialtyId equals specialty.IdSpecialty
-                       where application.Priority == 0
-                       group application by specialty.IdSpecialty into specialtyGroup
-                       select new ApplicantCountForSpecialityDto
-                       {
-                           SpecialtyId = specialtyGroup.Key,
-                           ApplicantCount = specialtyGroup.Count()
-                       })
-                       .ToList();
+        var request = AnalayzerService.GetCountApplicantToSpecialty();
         return Ok(request);
     }
 
@@ -104,18 +64,7 @@ public class RequestController(IRepository<Applicant> repositoryApplicant, IRepo
     [HttpGet("top-five-applicant-by-score")]
     public ActionResult<IEnumerable<TopFiveApplicantDto>> GetTopFiveApplicantToScore()
     {
-        var request = (from examResult in repositoryExamResult.GetAll()
-                       join applicant in repositoryApplicant.GetAll() on examResult.ApplicantId equals applicant.IdApplicant
-                       group examResult by applicant into applicantGroup
-                       let totalScore = applicantGroup.Sum(result => result.Score)
-                       orderby totalScore descending
-                       select new TopFiveApplicantDto
-                       {
-                           Applicant = applicantGroup.Key,
-                           TotalScore = totalScore
-                       })
-                       .Take(5)
-                       .ToList();
+        var request = AnalayzerService.GetTopFiveApplicantToScore();
         return Ok(request);
     }
 
@@ -128,26 +77,7 @@ public class RequestController(IRepository<Applicant> repositoryApplicant, IRepo
     [HttpGet("get-applicant-specialty-max-score")]
     public ActionResult<IEnumerable<ApplicantWithMaxScoreWithSpecialty>> GetApplicantAndSpecialtyMaxScore()
     {
-        var request = (from examResult in repositoryExamResult.GetAll()
-                       join applicant in repositoryApplicant.GetAll() on examResult.ApplicantId equals applicant.IdApplicant
-                       join application in repositoryApplication.GetAll() on applicant.IdApplicant equals application.ApplicantId
-                       join specialty in repositorySpecialty.GetAll() on application.SpecialtyId equals specialty.IdSpecialty
-                       group examResult by examResult.ExamId into examGroup
-                       let maxScore = examGroup.Max(result => result.Score)
-                       select new ApplicantWithMaxScoreWithSpecialty
-                       {
-                           Applicant = (from result in examGroup
-                                        where result.Score == maxScore
-                                        join applicant in repositoryApplicant.GetAll() on result.ApplicantId equals applicant.IdApplicant
-                                        select applicant).FirstOrDefault(),
-                           MaxScore = maxScore,
-                           SpecialtyId = (from result in examGroup
-                                          where result.Score == maxScore
-                                          join applicant in repositoryApplicant.GetAll() on result.ApplicantId equals applicant.IdApplicant
-                                          join application in repositoryApplication.GetAll() on applicant.IdApplicant equals application.ApplicantId
-                                          select application.SpecialtyId).FirstOrDefault()
-                       })
-                       .ToList();
+        var request = AnalayzerService.GetApplicantAndSpecialtyMaxScore();
 
         return Ok(request);
     }
